+++ TravelPlanner-Web/src/components/ui/checkbox.tsx
@@ -1,0 +1,30 @@
+/**
+ * TravelPlanner Web - Checkbox Component
+ *
+ * Checkbox form component.
+ */
+
+import * as React from "react";
+import { cn } from "@/lib/utils";
+
+export interface CheckboxProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {}
+
+const Checkbox = React.forwardRef<HTMLInputElement, CheckboxProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <input
+        ref={ref}
+        type="checkbox"
+        className={cn(
+          "h-4 w-4 rounded border-gray-300 text-blue-600 focus:ring-2 focus:ring-blue-600 focus:ring-offset-2 cursor-pointer",
+          className
+        )}
+        {...props}
+      />
+    );
+  }
+);
+Checkbox.displayName = "Checkbox";
+
+export { Checkbox };